# optimize allocations in redactStartedInformationCmd