<!--- If applicable, issue number goes here, e.g. GODRIVER-ABCD -->

## Summary
I suggest to optimize allocations in redactStartedInformationCmd function used in debug logging commands and command started events.
<!--- A summary of the changes proposed by this pull request. -->

## Background & Motivation
In one of our services 20% of allocations come from this function. We use only events for metrics. CPU consumption is not that large - only 1.5%, but I believe some part of GC CPU consumption that stands for more than 10% of total service CPU consumption also originates from these allocations.

Benchmark attached gives the following results without suggested changes:
```
cpu: Intel(R) Core(TM) i9-14900K
BenchmarkRedactStartedInformationCmd
BenchmarkRedactStartedInformationCmd/0
BenchmarkRedactStartedInformationCmd/0-32  		40476044	     70.50 ns/op	     112 B/op	       1 allocs/op
BenchmarkRedactStartedInformationCmd/1
BenchmarkRedactStartedInformationCmd/1-32  		 4045360	     296.6 ns/op	     320 B/op	       2 allocs/op
BenchmarkRedactStartedInformationCmd/5
BenchmarkRedactStartedInformationCmd/5-32  		 3126498	     428.6 ns/op	     320 B/op	       2 allocs/op
BenchmarkRedactStartedInformationCmd/10
BenchmarkRedactStartedInformationCmd/10-32 		 3266930	     471.1 ns/op	     320 B/op	       2 allocs/op
BenchmarkRedactStartedInformationCmd/100
BenchmarkRedactStartedInformationCmd/100-32         	  222031	      4748 ns/op	    2656 B/op	       5 allocs/op
BenchmarkRedactStartedInformationCmd/1000
BenchmarkRedactStartedInformationCmd/1000-32        	   26457	     62563 ns/op	   45280 B/op	      12 allocs/op
```
And with this changes:
```
BenchmarkRedactStartedInformationCmd
BenchmarkRedactStartedInformationCmd/0
BenchmarkRedactStartedInformationCmd/0-32  		27693193	     85.49 ns/op	     112 B/op	       1 allocs/op
BenchmarkRedactStartedInformationCmd/1
BenchmarkRedactStartedInformationCmd/1-32  		10653050	     138.6 ns/op	     112 B/op	       1 allocs/op
BenchmarkRedactStartedInformationCmd/5
BenchmarkRedactStartedInformationCmd/5-32  		 4048020	     266.7 ns/op	     144 B/op	       1 allocs/op
BenchmarkRedactStartedInformationCmd/10
BenchmarkRedactStartedInformationCmd/10-32 		 7370712	     363.6 ns/op	     192 B/op	       1 allocs/op
BenchmarkRedactStartedInformationCmd/100
BenchmarkRedactStartedInformationCmd/100-32         	  571992	      1924 ns/op	    1024 B/op	       1 allocs/op
BenchmarkRedactStartedInformationCmd/1000
BenchmarkRedactStartedInformationCmd/1000-32        	   62100	     21280 ns/op	   10240 B/op	       1 allocs/op
```

For some runs degradation of the case with 0 documentSequences is even bigger. I don't know how common is this case.